+++ alembic/env.py
@@ -1,6 +1,8 @@
+import asyncio
 from logging.config import fileConfig
 
-from sqlalchemy import engine_from_config, pool
+from sqlalchemy import Connection, engine_from_config, pool
+from sqlalchemy.ext.asyncio import async_engine_from_config
 
 from alembic import context
 from database.base import ProductionBase
@@ -26,12 +28,8 @@ target_metadata = ProductionBase.metadata
 # can be acquired:
 # my_important_option = config.get_main_option("my_important_option")
 database_url = get_database_settings().database_url
-# Replace async driver with sync driver for Alembic
-sync_database_url = database_url.replace("+asyncpg://", "://")
-# Replace docker service name with localhost for host-based Alembic runs
-sync_database_url = sync_database_url.replace("db:5432", "localhost:8432")
-print(f"Database URL: {sync_database_url}")
-config.set_main_option("sqlalchemy.url", sync_database_url)
+print(f"Database URL: {database_url}")
+config.set_main_option("sqlalchemy.url", database_url)
 # ... etc.
 
 
@@ -59,6 +57,29 @@ def run_migrations_offline() -> None:
         context.run_migrations()
 
 
+def do_run_migrations(connection: Connection) -> None:
+    context.configure(connection=connection, target_metadata=target_metadata)
+
+    with context.begin_transaction():
+        context.run_migrations()
+
+
+async def run_async_migrations() -> None:
+    """
+    In this scenario, we need to create an Engine
+    and associate a connection with the context.
+    """
+    connectable = async_engine_from_config(
+        config.get_section(config.config_ini_section, {}),
+        prefix="sqlalchemy.",
+        poolclass=pool.NullPool,
+    )
+    async with connectable.connect() as connection:
+        await connection.run_sync(do_run_migrations)
+
+    await connectable.dispose()
+
+
 def run_migrations_online() -> None:
     """Run migrations in 'online' mode.
 
@@ -66,20 +87,7 @@ def run_migrations_online() -> None:
     and associate a connection with the context.
 
     """
-    connectable = engine_from_config(
-        config.get_section(config.config_ini_section, {}),
-        prefix="sqlalchemy.",
-        poolclass=pool.NullPool,
-    )
-
-    with connectable.connect() as connection:
-        context.configure(
-            connection=connection,
-            target_metadata=target_metadata,
-        )
-
-        with context.begin_transaction():
-            context.run_migrations()
+    asyncio.run(run_async_migrations())
 
 
 if context.is_offline_mode():
