+++ app_factory/__init__.py
@@ -1,0 +1,5 @@
+"""
+Initialize the fastapi app instance for the application.
+
+With the necessary configurations and middlewares.
+"""
